Move logBook logic to a store
