+++ digital-guardbook-ui/src/models/logBook.ts
